refactor: honest file names for board-button, board-file-types, device-output, app-info
- board-button.ts held two read-helpers while the BoardButton type
  lives in types.ts; renamed to button-readers.ts, matching content.
- board-file-types.ts exports extensions and a predicate, no
  TypeScript types; renamed to board-file-formats.ts.
- device-output.ts's "device output" concept had no consumer; split
  into stub-speech.ts and stub-audio.ts along its two mock groups.
- app-info.ts exported one constant; folded APP_NAME into its only
  consumer, onboarding-dialog.tsx.
