--- src/app/app-info.ts
@@ -1,1 +1,0 @@
-export const APP_NAME = "AAC Board AI";
