+++ src/features/board/activation/button-activation.test.ts
@@ -1,4 +1,5 @@
-import { stubAudio, stubSpeech } from "@shared/testing/device-output";
+import { stubAudio } from "@shared/testing/stub-audio";
+import { stubSpeech } from "@shared/testing/stub-speech";
 import { beforeEach, describe, expect, test, vi } from "vitest";
 import type { UseMessagePlaybackReturn } from "../message/playback/use-message-playback";
 import type { MessagePart, UseMessageReturn } from "../message/use-message";
