+++ src/features/board/board-viewer.tsx
@@ -3,11 +3,11 @@ import Stack from "@mui/material/Stack";
 import type { Theme } from "@mui/material/styles";
 import useMediaQuery from "@mui/material/useMediaQuery";
 import { m } from "@paraglide/messages.js";
-import { useLanguage } from "@shared/language/use-language";
 import { useHighlightConfig } from "@shared/highlight/highlight-store";
+import { useLanguage } from "@shared/language/use-language";
 import { safeAreaInset } from "@shared/theme/safe-area";
 import { createButtonActivation } from "./activation/button-activation";
-import { getNavigationTargetId } from "./board-button";
+import { getNavigationTargetId } from "./button-readers";
 import { Grid, type GridItemProps } from "./grid/grid";
 import { useBoardKeyboard } from "./keyboard/use-board-keyboard";
 import { BackspaceButton } from "./message/backspace-button";
